revamp

+++ src/dao/Dao.sol
@@ -6,7 +6,7 @@ import {DaoLogic} from "./libraries/DaoLogic.sol";
 
 contract Dao {
     uint256 public quorum;
-    uint256 public votingDuration;
+    uint8 public votingDuration;
     uint256 public proposalCount;
     mapping(address => uint256) public membershipBalance;
     mapping(uint256 => mapping(address => bool)) private hasVoted;
